Bring login_controller up
Views and handlers and associated hierarchical names (e.g. newflow_login_path) had to change.
