+++ app/views/login/check_your_email.html.erb
