+++ app/views/login/login_form.html.erb
@@ -12,7 +12,7 @@
                     header: (I18n.t(:"login_signup_form.login_page_header")),
                     banners: @banners,
                     show_exit_icon: true) do %>
-            <% lev_form_for :login_form, url: newflow_login_path do |f| %>
+            <% lev_form_for :login_form, url: login_path do |f| %>
                 <% fh = NewflowFormHelper::Newflow.new(f: f, context: self, errors: @handler_result&.errors) %>
                     <div class="content social-section">
                         <div><%= I18n.t(:"login_signup_form.login_with") %></div>
@@ -20,14 +20,26 @@
                         <div class="social-buttons">
 
                             <div class="fb-share-button">
-                                <%= link_to newflow_auth_path(:facebooknewflow), class: 'facebook btn' do %>
+                                <%=
+                                    link_to(
+                                        newflow_auth_path(:facebooknewflow, origin: :login_form),
+                                        class: 'facebook btn',
+                                        data: { ga_category: 'Login', ga_action: 'Click', ga_label: 'Facebook' }
+                                    ) do
+                                %>
                                     <i class="social-icon fa fa-facebook"></i>
                                     <span>Facebook</span>
                                 <% end %>
                             </div>
 
                             <div class="google-share-button">
-                                <%= link_to newflow_auth_path(:googlenewflow), class: 'google btn' do %>
+                                <%=
+                                    link_to(
+                                        newflow_auth_path(:googlenewflow, origin: :login_form),
+                                        class: 'google btn',
+                                        data: { ga_category: 'Login', ga_action: 'Click', ga_label: 'Google' }
+                                    ) do
+                                %>
                                     <i class="social-icon fa fa-google"></i>
                                     <span>Google</span>
                                 <% end %>
@@ -53,46 +65,45 @@
                         <%= fh.text_field name: :email,
                                     value: email_prefill_value,
                                     placeholder: I18n.t(:"login_signup_form.email_placeholder"),
-                                    readonly: email_prefill_value.present?
+                                    autofocus: true
                         %>
-
                     </div>
 
                     <div class="content control-group">
-                        <%= label_tag :password, I18n.t(:"login_signup_form.password_label"),
-                                class: 'field-label required'
-                        %>
                         <div id="password-field-container">
-                            <%= fh.text_field name: :password,
-                                        type: :password,
-                                        placeholder: I18n.t(:"login_signup_form.password_label"),
-                                        autofocus: true
+                            <%= label_tag :password, I18n.t(:"login_signup_form.password_label"),
+                                    class: 'field-label required'
                             %>
                             <span id="password-show-hide-button">
-                                <span class="toggle-show-hide">SHOW</span>
+                                <span class="toggle-show-hide" data-ga_category='SHOW/HIDE password', data-ga_action='Click', data-ga_label='SHOW - Logon'>SHOW</span>
                                 <span class="toggle-show-hide" style="display: none">HIDE</span>
                             </span>
+                            <%= fh.text_field name: :password,
+                                        type: :password,
+                                        placeholder: I18n.t(:"login_signup_form.password_label")
+                            %>
                         </div>
                     </div>
 
                     <div class="content">
                         <%= f.submit I18n.t(:"login_signup_form.continue_button"),
                                     disable_with: I18n.t(:"login_signup_form.continue_button"),
-                                    class: 'primary'
+                                    class: 'primary',
+                                    data: { ga_category: 'Login', ga_action: 'Click', ga_label: 'Email' }
                         %>
                     </div>
 
                     <div class="content">
-                        <%# This is one difference from the login form...
-                             #  POST request to send_reset_password_email_path
-                             #  instead of GET request to forgot_password_form_path
-                        %>
                         <%=
                             link_to(
                                 I18n.t(:"login_signup_form.forgot_password"),
-                                send_reset_password_email_path,
+                                forgot_password_form_path,
                                 id: 'forgot-password-link',
-                                method: :post
+                                data: {
+                                    ga_category: 'Login',
+                                    ga_action: 'Click',
+                                    ga_label: 'Password Reset'
+                                }
                             )
                         %>
                     </div>
