+++ app/views/newflow/_tab_group.html.erb
@@ -1,5 +1,5 @@
 <div class="tab-group">
-    <%= link_to I18n.t(:"login_signup_form.log_in"), newflow_login_path,
+    <%= link_to I18n.t(:"login_signup_form.log_in"), login_path,
                     class: "tab login #{active_class}"
     %>
     <%= link_to I18n.t(:"login_signup_form.sign_up"), newflow_signup_path,
