+++ app/views/newflow/password_management/forgot_password_form.html.erb
@@ -14,7 +14,7 @@
                     )
                 %>
 
-                <%= render partial: 'newflow/back_arrow', locals: { path: newflow_login_path } %>
+                <%= render partial: 'newflow/back_arrow', locals: { path: login_path } %>
 
                 <div class="content control-group info-message">
                     <%= I18n.t(:"login_signup_form.reset_my_password_description") %>
