+++ app/views/newflow/social_auth/social_login_failed.html.erb
@@ -20,7 +20,7 @@
                             <%= f.submit(I18n.t(:"login_signup_form.send_me_instructions"),
                                                      id: 'send-instructions-button') %>
 
-                            <%= link_to newflow_login_path do %>
+                            <%= link_to login_path do %>
                                 <%= button_tag(I18n.t(:"login_signup_form.return"), type: nil, id: 'return-button') %>
                             <% end %>
                         </div>
