+++ spec/features/newflow/require_recent_sign_in_to_change_authentications_spec.rb
@@ -14,7 +14,7 @@ feature 'Require recent log in to change authentications', js: true do
 
   scenario 'adding Facebook' do
     visit '/'
-    newflow_log_in_user(email_value, 'password')
+    log_in_user(email_value, 'password')
 
     expect(page.current_path).to eq(profile_newflow_path)
 
@@ -45,7 +45,7 @@ feature 'Require recent log in to change authentications', js: true do
   scenario 'changing the password' do
     with_forgery_protection do
       visit '/'
-      newflow_log_in_user(email_value, 'password')
+      log_in_user(email_value, 'password')
       expect(page).to have_no_missing_translations
 
       Timecop.freeze(Time.now + RequireRecentSignin::REAUTHENTICATE_AFTER) do
@@ -82,7 +82,7 @@ feature 'Require recent log in to change authentications', js: true do
   end
 
   # scenario 'bad password on reauthentication' do
-  #   newflow_log_in_user(email_value, 'password')
+  #   log_in_user(email_value, 'password')
   #
   #   Timecop.freeze(Time.now + RequireRecentSignin::REAUTHENTICATE_AFTER) do
   #     visit profile_newflow_path
@@ -110,7 +110,7 @@ feature 'Require recent log in to change authentications', js: true do
     with_forgery_protection do
       FactoryBot.create :authentication, user: user, provider: 'twitter'
 
-      newflow_log_in_user(email_value, 'password')
+      log_in_user(email_value, 'password')
 
       expect(page).to have_no_missing_translations
       Timecop.freeze(Time.now + RequireRecentSignin::REAUTHENTICATE_AFTER) do
